requirements.txt
requests~=2.32.2
python-dotenv~=1.0.1
loguru~=0.5.3
prettytable~=3.5.0